PP-5757 Log the correct id when we process a Smartpay notification
We were logging the "pspReference", but this is a unique reference for
the operation, rather than the Smartpay's reference for the charge.
This is unhelpful as we don't have an easy way to relate it back to the
charge. Instead, log the "originalReference" which is Smartpay's
reference for the charge.