PP-5757: Log the SmartpayNotification properly
We get a lot of Smartpay notifications for which there aren't any corresponding
charges. Over the period of 21-25th October there were 469-622 such events each
day. Log the details of the notification properly so we can get more
information as to whether we need to take any action on this.