PP-5757 Don't log error when notification has already been processed
We log an error if we fail to transition a charge when processing a
notification. We've seen a number of errors logged when we try to
transition the charge into a state it's already in. In this case we've
already processed the notification so it's fine to do nothing and log
nothing.

We're still logging an error if we can't transition to a different
state, however we should review if we see errors like this whether they
are actually errors.