PP-5757 Don't log attempt to cancel charge in not allowed state as error
For attempts to cancel a charge that isn't in a legal state, we will
send an error response which is either relayed to the service in the
public API response, or handled by frontend. This is not an error, so
log as INFO.

For an attempt to expire a charge not in a valid state, this ideally
shouldn't happen as we should only attempt to expire charges in valid
states, but there could be some race involved here. In which case it
might be indicative that our expiry process isn't ideal but not an
error we need to immediately act on. So log this at WARN level.