PP-5757 Improve 3DS Authorisation response
We were doing as follows:
- If the authorisation status was AUTHORISED or AUTH_3DS_READY we
returned a 200 response frontend.
- If the authorisation status was REJECTED or ERROR we returned a 400
response to frontend.
- If the authorisation status was anything else, we returned a 500 to
frontend, causing it to show the technical difficulties page.

However, this did not account for the CANCELLED status which is a valid
outcome for Worldpay 3DS authorisations, which we have observed come
back.

Improve this by still returning a 200 for AUTHORISED or AUTH_3DS_READY
statuses, and a 500 only for the EXCEPTION status, and a 400 for any
other status.