PP-5757 Allow CAPTURE SUBMITTED to CAPTURE ERROR transition
We can receive Smartpay notifications informing us that a capture didn't
succeed. Currently, as this state transition is disallowed we do nothing
to the charge when we receive this notification. The charge therefore
remains stuck in the "CAPTURE SUBMITTED" state.

This also means that a ticket is not created in Zendesk to inform us of
the capture error, which should be the case.