PP-5757 Increase ePDQ capture timeout even more
We've seen great results by increasing this from 1s to 3s, but there are
requests that come close to 3s to respond.

Increase the timeout to 20s in line with the other requests
(auth/cancel/refund) to give it lots of headroom. This should mean that
we should rarely or never have to retry requests, meaning our capture
queue size won't always be huge and is more indicative of the volume of
payments rather that just being full of ePDQ payments retrying.

This will also provide a better experience for services, as they will be
able to refund immediately in most cases rather than having to wait a
day or 2 until we get a notification from ePQD to tell us the payment
was successfully captured.