PP-5812 Add payment provider to StateTransitionService log line
To enable our payment journey SLI we need to know the payment
provider when emitting events so that we can exclude `sandbox`. This is
being added as a structured argument.

Also take the opportunity to add the `to_state` and `from_state` as
structured arguments to the logger to make the implementation of the sli
query in Splunk simpler.