PP-5757 Allow transition from CREATED to USER CANCELLED
There is an edge case where a secure token can be used but the page is
reloaded before the redirect to the enter card details page. This means
that the token has been used but the payment hasn't been transitioned to
ENTERING CARD DETAILS.

In this case, if the user re-visits the secure page and doesn't have the
session key in their cookie an error page will be shown, which will
attempt to cancel the payment when a link is clicked to return to the
service.

Allow this state transition.