PP-5709: add shopperIPAddress to authorisation request (Worldpay)
Adding shopperIPAddress attribute to authorisation request can be
used by Worldpay for fraud checks.
It's an IP address which has been used by payer.
We want to add it in all cases (not only for 3ds payments).

Changes:
* update to AuthCardDetails to include ip address
* update to WorldpayAuthoriseOrderTemplate.xml with a change to the
condition for displaying session element
when it's 3ds payment or we have information about IP address
* update to tests (remove duplicate one)