PP-5809 Log charge state conflicts at info level
If Connector is asked by a client to perform a change that violates our
state model this shouldn't be logged as an error, it is for the client
to decide in that particular context whether it should be.
`conflictErrorResponse` is mostly called when trying to perform delayed
capture on a charge that isn't in `AWAITING CAPTURE REQUEST`.