PP-5809 Lower log level in NotificationsResource
Lower log level when rejecting notifications within the
`NotificationsResource` to `info` since it doesn't have the full context
to know whether its an error. The method
`handleNotifications` which is called prior to the log line being changed in this
commit has dedicated log lines at various levels to cover
the different rejection scenarios.