PP-6075 Better rationalise ChargeEntity constructors
It was unclear what the difference between the various ChargeEntity
constructors were. Make this clearer by:

- Moving `ChargeEntityFixture` to the same package as `ChargeEntity` in
test code and making the constructor used only by `ChargeEntity`
package-private.

- Adding static factory methods for constructing `ChargeEntity` to make
it immediately clear that one is for creating a web payment and the
other is for telephone payments.

- Reordering the parameters of the constructor used by telephone
payments to make it easier to see what the difference is between this
constructor and the constructor used for non-telephone payments. Also
remove `status`, `source` and `language` parameters as these are always
given static values.