PP-6075 Remove migration to backfill moto column
This migration was very long-running and caused the connector database
to become overloaded due to the large transaction size, severely
affecting latency of calls to connector. It was therefore aborted.

Remove the migration so it will not be re-run if we attempt to migrate
the database again. A few different solutions are being explored to set
the default value including upgrading to Postgres 11 which will set
default values the next time the column is read, or running the update
query in batches.