PP-6072 Add DAO method to search for gateway accounts
Add method to search gateway accounts by multiple parameters. Currently
this is just by gateway account ids and whether MOTO is enabled or not.
The is being added so we can filter gateway accounts in toolbox, and is
intended to be extended in future with additional search parameters.

Add GatewayAccountSearchParams which is the query parameters bean which
will be populated by the resource. This follows the pattern used for
searching in ledger.

This new DAO method will be used by the GET `/v1/api/accounts` search
resource, and will replace the existing `list` method in the DAO that
takes a list of gateway account ids as a parameter.