PP-6186: Cancel charge with gateway if it's in a terminable state
The scenario here is the attempt to cancel an ePDQ charge by a user or via the
API. With ePDQ integration, some services configure their payments to be
"one-step", that is, a charge is authorised and captured in one step. (GOV.UK
Pay assumes "two-step", that is, separate API calls have to be made for
authorisation and capturing a charge.)

The result of this is that when a cancel is attempted, the charge may already
have been captured, which is a non-terminable state. This commit adds a check
to see if the charge is in a terminable state before canceling.

At present, the only implementation of the PaymentProvider that implements the
`queryPaymentStatus` method is the EpdqPaymentProvider. We have therefore
amended the `QueryService.isTerminableWithGateway` method to return `true` if
the implementation of PaymentProvider throws an UnsupportedOperationException.
This is the equivalent of saying we assume the charge can be canceled if the
PSP is not epdq which is a reasonable assumption because the only complaints we get
are services which use ePDQ as their PSP.