PP-6186 Attempt to cancel payments with gateway when we do not query status
It seems there was possibly a slight regression a while ago for payment
cancellation and expiry.

Previously, we would always attempt to cancel payments that were in
certain internal states with the gateway and just swallow the error if
this failed and just transition the payment to cancelled locally.

We changed this to attempt to cancel a greater range of internal states
with the gateway, but query first to see whether this was possible. This
was done mainly in order to stop certain payments being stuck in an
authorised state indefinitely with ePDQ.

However, querying the status with the gateway is not implemented for
Smartpay or Stripe, and in these cases we will no longer attempt to
cancel with the gateway for any payments.

Modify this so that if we haven't queried the status with the gateway
due to this not being implemented, we will always attempt to cancel the
payment with the gateway when we expire or cancel a payment and swallow
the error if this is not possible. We will attempt to cancel charges
with the gateway if they are in the following states:
`AUTHORISATION_3DS_REQUIRED`
`AUTHORISATION_3DS_READY`
`AUTHORISATION_SUCCESS`
`AWAITING_CAPTURE_REQUEST`

Also modify what we log if cancellation with the gateway fails to make
it easier to search for.