PP-6189 Use gateway status if terminal instead of expiring
If a charge is picked up by the expiry sweep has a status that indicates
the gateway knows about it, then for ePDQ and Worldpay payments:

- we get the status of the charge with the gateway and map it to a
ChargeStatus.
- if the mapped status is terminal (CAPTURED, AUTHORISATION REJECTED)
then we don't expire and instead attempt to update the status of the
charge to match the gateway status.
- if the mapped status is not terminal, we attempt to cancel the charge
with the gateway and then expire it.

For other gateways, we do not check with the gateway and instead always
try to cancel with the gateway if the charge has been authorised
successfully.