PP-6189 Allow for normal state transition to terminal state on expiry
If when we query the payment gateway before expiry, and it tells us that
the charge maps to a terminal state that we can normally transition to
without "forcing" it, then transition to that state.

This covers that case where 3DS authorisation has been rejected for an
ePDQ payment, but the user was never returned to our pages from the ePQQ
3DS page so we never knew about it until this point.

Also, when we update the status of the charge rather than expiring it,
treat this as an expiry success in the response returned from the
expiry sweep.