PP-6238 Enable prometheus metrics
If `PROMETHEUS_ENABLED` is `true` add a `DropwizardExports` to process
the default `MetricsRegistry` into the format required by Prometheus and
add it to the default Prometheus `CollectorRegistry`. To provide an
end-point which Prometheus can scrape, add a `MetricsServlet` exposed on
a configurable path and add serve it via the admin port.

The default settings are to not enable Prometheus.