PP-6194 Add service for ePDQ gateway cleanup
Add service for the recurring job to cleanup ePDQ charges that are in
"AUTHORISATION ERROR", "AUTHORISATION UNEXPECTED ERROR" or
"AUTHORISATION TIMEOUT" statuses.

This job is required as when we enter into one of these statuses we are
unsure whether the charge was actually authorised with the gateway. If
it was authorised, we need to cancel it on the gateway to ensure that
funds are released in the paying users' bank account.

The job will move the charge into one of three statuses when it has
successfully handled it:
"AUTHORISATION ERROR CANCELLED" - the charge was authorised on the
gateway but has now been cancelled.
"AUTHORISATION ERROR REJECTED" - the authorisation was rejected on the
gateway and no action needed to be taken to clean up.
"AUTHORISATION ERROR CHARGE MISSING" - the charge was not found on the
gateway, most likely because the error was before the gateway processed
the authorisation.

Map ePDQ status code "0" to "AUTHORISATION ERROR" status when querying
the charge status, as this represents an authorisation response that we
would have handled by moving the charge into this state.