PP-6194 Return whether charge found when querying gateway
Want to be able to differentiate between the charge not being found on
the gateway and it being in a state that we don't have a mapping for.

To achieve this, move mapping of the charge status into
WorldpayPaymentProvider and EpdqPaymentProvider and accept the
BaseInquiryResponse when constructing a ChargeQueryResponse. Determine
if a charge was found by looking at whether the BaseInquiryResponse has
a `transactionId` set.

BaseInquiryResponse previously had a `getLastEvent` method, but this is
Worldpay-specific and not used by the interface, so repurpose this
interface across the gateway query responses and make EpdqQueryResponse
implement it.

Add some more tests for querying the charge status and fix
`it/contract/EpdqPaymentProviderTest.java`.