PP-6194 Clean up WorldpayPaymentProviderTest somewhat
Remove base class WorldpayBasePaymentProviderTest, as there was only
a single subclass of this and move code into WorldpayPaymentProviderTest.

Make more clear which tests use a real GatewayClient (with a mocked
client) and a mocked GatewayClient. Ideally there wouldn't be this mix
of tests, but I don't want to dig too deep into changing this right now.

Replace some xml responses that existed in the test class with uses of
the test Worldpay response templates.