PP-6194 Add limit for gateway cleanup job
Take a mandatory query param, "limit", for the
v1/tasks/gateway-cleanup-sweep endpoint, which limits the number of
charges that are processed by the job.

The intention of this is to have a safe limit to ensure that we won't
end up sending a large number of requests to ePDQ in an out of the
ordinary situation that may hit a rate limit.