PP-6224: Don't expunge epdq charges in select states
Don't expunge if in AUTHORISATION ERROR, AUTHORISATION TIMEOUT and
AUTHORISATION UNEXPECTED ERROR states.
The reason we're doing this is because we don't trust that epdq charges in
these states match the states at the epdq end. On the epdq side, these charges
can be:
  - missing
  - captured
  - authorised
  - rejected

By making these not expungable, the EpdqAuthorisationErrorGatewayCleanupService
will be able to pick these charges up and clean them up.