DRAFT Validate External Metadata on ChargeEntity
Validate the `ExternalMetadata` when creating a new `ChargeEntity`
to prevent invalid external meta data being created from within
Connector (e.g. not via an api request). This can pose a problem for
`Selfservice` when it attempts to display a charge with invalid meta
data.

Previous attempt used the `@Valid` annotation but this prevented existing
charges with invalid metadata being updated (the validation was called upon
any JPA update and persist operation). Furthermore the error message reported was
unclear that the exception was being thrown by JPA and not the API
layer.

This approach permits updating existing charges which have metadata
considered to be invalid by the latest standards of `ValidExternalMetadata`
which is important since this standard can evolve over time. It
prevents new `ChargeEntities` from being created if the external
metadata does not conform to the latest `ValidExternalMetadata`
standard. The exception message should provide a clear indication what
is wrong.