PP-6259 Rename Auth3dsDetailsEntity to Auth3dsRequiredEntity
Rename Auth3dsDetailsEntity to Auth3dsRequiredEntity to better
indicate that it represents the data that’s used when we’re
required to do 3D Secure (e.g. a paRequest) and to avoid
confusion with Auth3dsDetails, which represents the result of 3D
Secure authentication (containing e.g. a paResponse).

with @SandorArpa