PP-6194 Handle charges already cancelled with gateway
There are some ePDQ charges that are in an `AUTHORISATION ERROR` on our
end but are `USER CANCELLED` on the gateway. This is because we have
already cancelled these charges with the gateway manually using the
discrepancy checker.

Handle these charges by simply moving them to the `AUTHORISATION ERROR
CANCELLED` state.