PP-6194 Allow force transition to auth error and rejected states
The charge could be in a terminal state corresponding to AUTHORISATION
ERROR or AUTHORISATION REJECTED with the gateway. Currently when we come
across charges in these states as part of the expiry job we don't expire
them and leave them to be picked up continuously by the job.

Instead we should transition them to the terminal state they are in on
the gateway.