PP-6194 Fix charge status transition not persisted for cleanup job
The update to the charge status was not being persisted because the
ChargeEntity was not part of the transaction opened for the
ChargeService method.

Switch to using the ChargeService method that takes the charge external
id as a parameter so that the ChargeEntity is loaded into the transaction
that it is updated in.