PP-6194 Fix charge status transition not persisted for expiry job
When updating the charge status to match the gateway status, the change
was not being persisted because the ChargeEntity being modified was not
part of the current transaction.

Ensure we are loading the ChargeEntity into the transaction and add a
test for this.

Improve the tests for the expiry service to ensure we are persisting
the status change.