PP-6259 Reduce maximum number of 3D Secure attempts from 3 to 1
Reduce the maximum number of times a user is allowed to attempt
3D Secure from 3 to 1 for gateways we allow retries (only
Worldpay). After their 1 attempt, if the gateway asks us to do
3DS again, we’ll consider the payment to be AUTHORISATION REJECTED.

We’re doing this because the majority of times when the user
retries 3DS results in Worldpay sending an error code 7 with the
message “verification of PaRes failed” when we tried to authorise
the second result. This means we’re making users do 3DS twice and
they’re still not able to pay, which is a worse experience than
when we just gave them an error after the first attempt.

with @sandorarpa