BAU: Stripe refund hanlder sets transfer id for gateway transaction
We currently set the ID of the refund against the platform for the
refund `gateway_transaction_id`. The actual refund relative to the
Connect account is the transfer from the Connect account to the
platform.

Store this ID so we can use to relate refund transactions and GOV.UK Pay
refunds during reconcilation.