PP-6477: simplify historic charge calculation
For general happy paths, given a charge on this code path is expunged we
can guarantee that the payment status is terminal.

Given this, we should be recalculating refund-ability if something on
the refund could have changed but we don't need to include the charge
pending states.

Include a definition for refund states that can be mutated (given either
a successful or failed refund) and recalculate amount available given
those cases.