PP-6470 Log an error if no payments/refunds found for payout
Log an error if the BalanceTransactions retrieved for a payout does not
contain any payments or refunds.

We want to ensure that if the Stripe API changes in a way that we do not
notice when upgrading the SDK, we are alerted by Sentry if we stop
successfully retrieving transactions for the payout.

In this case, we also don't want to mark the message as processed. So we
can fix and retry if necessary.

Also change the log level to error when we process a BalanceTransaction
of an unexpected type so these reach Sentry.