PP-6475: migration to add swap table for emitted events
An issue with expunging emitted events from Connector was recently
addressed in https://github.com/alphagov/pay-connector/pull/2340. This
issue means that the `emitted_events` table has a large number of rows
that need to be safely removed.

The first step in the process outlined in this ticket is to create an
identical table that live Connector nodes can use to track emitted
events for in-flight payments.