PP-6475: point connector to emitted events swap table
Depends on https://github.com/alphagov/pay-connector/pull/2346. Moves
all live Connector nodes to point to swap table to allow `ACCESS
EXCLUSIVE` on the `emitted_events` table.

This pull request will be reverted once work has been completed on the
original table.

This is the second step in process outlined on this ticket.