PP-6444 Do not stop shared postgres container during IT
`PostgresDockerRule` effectively has a singleton postgres container which
is shared between  all IT tests using that rule. `HealthCheckResourceIT` was
calling the public stop method which in turn stopped the container and
set it to null after other tests had initialised and proceeded beyond
the container null check (which would start the postgres container if
necessary). We saw this on Concourse and perhaps the different
environment effected the running order and or timing.

Rather than use an app with postgres and then stop postgres to get the
desired result from the health check endpoint, just run the test with a
standard DropwizardAppRule which will not have postgres or sqs.

with @mrlumbu