Refactor iteration into separate iterator class
The iteration over batches of events is somewhat complicated and entangled
with other logic. By pulling it out into a separate iterator class we can
more clearly see which parts of the logic have to do with iteration, counters,
ranges etc.

The remaining business logic of actually emitting the events is then much clearer to see.

Some small bugs/issues were found during this refactoring due to the mockist
style of testing. These will be cleaned up in a subsequent commit.