Alter tests to reflect actual behaviour of chargeService
chargeService throws an exception if charge not found

this highlights that coupling with the mock default behaviour of returning null

This in turn brings to the surface that each event triggers a full backfill
for the charge entity and associated refunds. So for each event there will be
two events emitted. This seems not ideal behaviour but at least the tests
show the actual behaviour now.

I'm not keen on these heavily mocked tests. IMHO we should hit the DB for service
layer tests especially ones like this which rely on DB logic.