PP-6671 ePDQ 3DS2: send default browser data if provided too long
When sending an authorisation request for a 3DS2 payment to ePDQ,
don’t exceed the character limits of the browserAcceptHeader,
browserLanguage and browserUserAgent parameters of 2048, 8 and
2048 characters respectively because ePDQ seems to enforce these
limits for browserAcceptHeader and browserUserAgent, causing
failures (the limit doesn’t seem to be enforced for
browserLanguage but let’s enforce the limit there too in case it
starts being enforced in the future).

Where the provided data exceeds the limit, send the existing
default value we use if the data isn’t provided i.e. */* for
browserAcceptHeader, en-GB or cy for browserLanguage (depending on
the language of the payment) and Mozilla/5.0 for browserUserAgent.

For the browserAcceptHeader, we could try to do something fancy
like remove content types with lower q values until we get under
limit but this is a lot of effort for an edge case and would
violate ePDQ’s requirement to send the exact value of the HTTP
Accept header. Similarly, we could trim subtags from the
browserLanguage but again this would be a lot of effort for an
edge case.

Don’t enforce the limits for the HTTP_ACCEPT and HTTP_USER_AGENT
parameters (sent for both 3DS1 and 3DS2 authorisations), which
normally match the browserAcceptHeader and browserUserAgent
parameters respectively, because the limits don’t seem to be
enforced there. While this creates an inconsistency, ePDQ’s docs
suggest that HTTP_ACCEPT and HTTP_USER_AGENT are used to determine
if the user’s browser is compatible with issuer identification
system, so we want them to be accurate rather than defaults.