PP-6671 Do not send ePDQ 3DS2 address data that exceeds limits
Do not send ECOM_BILLTO_POSTAL_STREET_LINE1,
ECOM_BILLTO_POSTAL_STREET_LINE2, ECOM_BILLTO_POSTAL_CITY,
ECOM_BILLTO_POSTAL_POSTALCODE and ECOM_BILLTO_POSTAL_COUNTRYCODE
parameters to ePDQ when authorising a 3D Secure 2 payment if they
exceed ePDQ’s prescribed limits of 35, 35, 25, 10 and 2 characters
respectively because some of our own limits our longer than
these. Don’t send a parameter at all if its limit is exceeded
(these parameters are optional but recommended).

No changes to the equivalent OWNERADDRESS, OWNERTOWN, OWNERZIP and
OWNERCTY parameters sent to ePDQ for all requests where the user
supplies an address because the limits don’t seem to be enforced
there and we really don’t want to mess with the user-supplied
address we send unless we have to have to. (An exception is the
OWNERZIP parameter, which holds the postcode and seems to have its
10-character limit enforced, but this already matches our own
10-character limit for the postcode.)

Similarly, the CN (cardholder name) parameter we already send for
all requests (and is mandated by ePDQ for 3DS2) theoretically has
a 35 character limit but it doesn’t seem to be enforced (and
neither does the rule that quotes must be avoided) so no changes
there.