PP-6712 Don’t use EpdqTemplateData for maintenance requests
When building a payload to cancel, capture or refund an ePDQ
payment, don’t use EpdqTemplateData. Instead, set the data
directly on the EpdqPayloadDefinition subclass object.

Push some stuff down from EpdqPayloadDefinitionForMaintenanceOrder
to the appropriate subclass to better reflect reality. For example,
a refund is the only maintenance order that can have an amount, so
only it needs an amount field. Similarly, a cancel is the only
maintenance order that either has a PAYID or an ORDERID, so only it
needs the logic for dealing with this. After pushing  these things
down, EpdqPayloadDefinitionForMaintenanceOrder isn’t really pulling
its weight, so delete it and have its subclasses directly extend
EpdqPayloadDefinition instead.