PP-JFDI: Stripe auth handles non card errors
POST to Stripe authorisation will reject with `4xx` for both card
valid card declines and corrupt requests. All of the codes listed in
https://stripe.com/docs/api/errors have never fired other than
`card_error` and `validation_error` when there is a problem with the
Platform accounts.

By default we should therefore have immediate alerts around anything
that isn't being treated as a `card_error`, this cases will usually
require us to intervene.