PP-6353: Log payment transition to hosted graphite
We want to be able to detect anomalous drops in successful live authorisations.
It is envisaged that we can do this by noticing trends in authorisation errors
and successes. However it is currently hard to see this in the way we send data
to hosted graphite (HG). For example, currently we have to drill into
`aggregates.production.connector.gateway-operations.worldpay.live.gatewa-account-id`
(which is fine) then into `authorise/authorise3ds`,
`without-billing-address/with-billing-address` which makes it difficult to
graph the data we want. This PR sends the states a payment goes into which will
help us graph the trends more easily. This is similar to what we send to splunk.