PP-6353 specify metric units
According to https://github.com/dropwizard/metrics/issues/946 the default precision for
rate metrics is milliseconds which means that the accuracy of the metrics can be lost due
to the precision of the numbers reporter. Rates measured in minutes is a more management granularity.

Also specify the duration unit as milliseconds (for latencies etc) which is the default
but this makes it explicit.